modules/kafka-consumer/requirements.txt
certifi==2020.6.20
grpcio==1.33.2
grpcio-tools==1.33.2
kafka==1.3.5
kafka-python==2.0.2
protobuf==3.13.0
six==1.15.0
psycopg2-binary==2.8.5